fixes to make it work + git ignore.

+++ cadlib/visualize.py
@@ -61,9 +61,9 @@ def create_by_extrude(extrude_op: Extrude):
 
 def create_profile_face(profile: Profile, sketch_plane: CoordSystem):
     """create a face from a sketch profile and the sketch plane"""
-    origin = gp_Pnt(*sketch_plane.origin)
-    normal = gp_Dir(*sketch_plane.normal)
-    x_axis = gp_Dir(*sketch_plane.x_axis)
+    origin = gp_Pnt(*sketch_plane.origin.astype(float))
+    normal = gp_Dir(*sketch_plane.normal.astype(float))
+    x_axis = gp_Dir(*sketch_plane.x_axis.astype(float))
     gp_face = gp_Pln(gp_Ax3(origin, normal, x_axis))
 
     all_loops = [create_loop_3d(loop, sketch_plane) for loop in profile.children]
